+++ src/render.rs
@@ -361,110 +361,6 @@ impl Render {
         })
     }
 
-    pub fn render(&mut self, world: &World) {
-        // Debug rendering
-        if self.render_colliders {
-            world.run_with_data(Self::draw_bodies, self);
-            world.run_with_data(Self::draw_one_sensors, self);
-            world.run_with_data(Self::draw_beams, self);
-        }
-    }
-
-    fn draw_one_sensors(
-        &mut self,
-        pos: View<Transform>,
-        sens_tag: View<OneSensorTag>,
-    ) {
-        for (tf, tag) in (&pos, &sens_tag).iter() {
-            let color = if tag.col.is_some() {
-                Color::new(0.00, 0.93, 0.80, 1.00)
-            } else {
-                GREEN
-            };
-
-            match tag.shape {
-                ColliderTy::Box { width, height } => draw_rectangle_lines_ex(
-                    tf.pos.x,
-                    tf.pos.y,
-                    width,
-                    height,
-                    1.0,
-                    DrawRectangleParams {
-                        offset: vec2(0.5, 0.5),
-                        rotation: tf.angle,
-                        color,
-                    },
-                ),
-                ColliderTy::Circle { radius } => draw_circle_lines(
-                    tf.pos.x,
-                    tf.pos.y,
-                    radius,
-                    1.0,
-                    color,
-                ),
-            }
-        }
-    }
-
-    fn draw_beams(
-        &mut self,
-        pos: View<Transform>,
-        beam_tag: View<BeamTag>,
-    ) {
-        // TODO: draw collision count
-        for (tf, tag) in (&pos, &beam_tag).iter() {
-            let color = GREEN;
-
-            draw_rectangle_lines_ex(
-                tf.pos.x,
-                tf.pos.y,
-                tag.length,
-                tag.width,
-                1.0,
-                DrawRectangleParams {
-                    offset: vec2(0.0, 0.5),
-                    rotation: tf.angle,
-                    color,
-                },
-            );
-        }
-    }
-
-    fn draw_bodies(
-        &mut self,
-        pos: View<Transform>,
-        body_tag: View<BodyTag>,
-    ) {
-        for (tf, tag) in (&pos, &body_tag).iter() {
-            let color = match tag.kind() {
-                BodyKind::Static => DARKBLUE,
-                BodyKind::Dynamic => RED,
-                BodyKind::Kinematic => YELLOW,
-            };
-
-            match tag.shape() {
-                ColliderTy::Box { width, height } => draw_rectangle_ex(
-                    tf.pos.x,
-                    tf.pos.y,
-                    *width,
-                    *height,
-                    DrawRectangleParams {
-                        // offset: Vec2::ZERO,
-                        offset: vec2(0.5, 0.5),
-                        rotation: tf.angle,
-                        color,
-                    },
-                ),
-                ColliderTy::Circle { radius } => draw_circle(
-                    tf.pos.x,
-                    tf.pos.y,
-                    *radius,
-                    color,
-                ),
-            }
-        }
-    }
-
     pub fn render_ui(
         &mut self,
         state: AppState,
